+++ src/modules/common/infrastructure/persistence/typeorm/typeorm-repository.ts
@@ -1,7 +1,7 @@
 import { Document, FindManyOptions } from 'typeorm';
 
 import { Criteria } from '@common/domain/criteria/criteria';
-import { TypeormCriteriaConverter } from '@common/infrastructure/persistence/typeorm/typeorm-criteria-converter';
+import { TypeormCriteriaConverter } from '@common/infrastructure/persistence/typeorm/criteria/typeorm-criteria-converter';
 
 export abstract class TypeormRepository<T extends Document> extends TypeormCriteriaConverter<T> {
   protected getQueryByCriteria(criteria: Criteria): FindManyOptions<T> {
